+++ Verizon/src/main/java/com/mopub/mobileads/VerizonAdapterConfiguration.java
@@ -5,9 +5,7 @@ import android.content.Context;
 import android.os.Handler;
 import android.os.Looper;
 import android.text.TextUtils;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
+import android.util.Base64;
 
 import com.mopub.common.BaseAdapterConfiguration;
 import com.mopub.common.OnNetworkInitializationFinishedListener;
@@ -22,8 +20,17 @@ import com.verizon.ads.VASAds;
 import com.verizon.ads.edition.StandardEdition;
 import com.verizon.ads.utils.ThreadUtils;
 
-import java.util.Map;
+import org.json.JSONException;
+import org.json.JSONObject;
 
+import java.io.ByteArrayOutputStream;
+import java.util.Map;
+import java.util.zip.DeflaterOutputStream;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CUSTOM_WITH_THROWABLE;
 import static com.verizon.ads.VASAds.ERROR_AD_REQUEST_FAILED;
 import static com.verizon.ads.VASAds.ERROR_AD_REQUEST_TIMED_OUT;
 import static com.verizon.ads.VASAds.ERROR_NO_FILL;
@@ -33,6 +40,11 @@ public class VerizonAdapterConfiguration extends BaseAdapterConfiguration {
     private static final String ADAPTER_VERSION = BuildConfig.VERSION_NAME;
     private static final String MOPUB_NETWORK_NAME = BuildConfig.NETWORK_NAME;
     private static final Handler handler = new Handler(Looper.getMainLooper());
+    private static final String VERIZON_ADS_DOMAIN = "com.verizon.ads";
+    private static final String EDITION_NAME_KEY = "editionName";
+    private static final String EDITION_VERSION_KEY = "editionVersion";
+    private static final String BIDDING_TOKEN_VERSION = "1.0";
+    private static String biddingToken = null;
 
     public static final String MEDIATOR_ID = "MoPubVAS-" + ADAPTER_VERSION;
     public static final String SERVER_EXTRAS_AD_CONTENT_KEY = "adm";
@@ -49,7 +61,12 @@ public class VerizonAdapterConfiguration extends BaseAdapterConfiguration {
     @Nullable
     @Override
     public String getBiddingToken(@NonNull Context context) {
-        return VASAds.getBiddingToken(context, VASAds.getRequestMetadata());
+
+        if (biddingToken == null) {
+            String uncompressedToken = getBiddingToken();
+            biddingToken = compressAndEncodeBiddingToken(uncompressedToken);
+        }
+        return biddingToken;
     }
 
     @NonNull
@@ -157,5 +174,50 @@ public class VerizonAdapterConfiguration extends BaseAdapterConfiguration {
             default:
                 return NativeErrorCode.NETWORK_INVALID_STATE;
         }
+    }
+
+
+    private String getBiddingToken() {
+
+        JSONObject biddingTokenJSON = new JSONObject();
+        JSONObject envJSON = new JSONObject();
+        JSONObject sdkInfoJSON = new JSONObject();
+        try {
+            String editionName = Configuration.getString(VERIZON_ADS_DOMAIN, EDITION_NAME_KEY, null);
+            String editionVersion = Configuration.getString(VERIZON_ADS_DOMAIN, EDITION_VERSION_KEY, null);
+            if ((editionName != null) && (editionVersion != null)) {
+                sdkInfoJSON.put("editionId", String.format("%s-%s", editionName, editionVersion));
+            }
+            sdkInfoJSON.put("version", BIDDING_TOKEN_VERSION);
+
+            envJSON.put("sdkInfo", sdkInfoJSON);
+            biddingTokenJSON.put("env", envJSON);
+            return biddingTokenJSON.toString();
+        } catch (JSONException e) {
+            MoPubLog.log(getMoPubNetworkName(), CUSTOM_WITH_THROWABLE, "Unable to get bidding token.", e);
+        }
+
+        return null;
+    }
+
+
+    private String compressAndEncodeBiddingToken(final String stringToCompress) {
+
+        if (TextUtils.isEmpty(stringToCompress)) {
+            return null;
+        }
+
+        ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
+        DeflaterOutputStream deflaterOutputStream = new DeflaterOutputStream(byteArrayOutputStream);
+        try {
+            deflaterOutputStream.write(stringToCompress.getBytes());
+            deflaterOutputStream.flush();
+            deflaterOutputStream.close();
+            return Base64.encodeToString(byteArrayOutputStream.toByteArray(), Base64.DEFAULT);
+        } catch (Exception e) {
+            MoPubLog.log(getMoPubNetworkName(), CUSTOM_WITH_THROWABLE, "Unable to compress bidding token.", e);
+        }
+
+        return null;
     }
 }